Set up tests for dflowfm
